backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
Pillow
pytesseract
requests
zai-sdk
pypdf

